requirements.txt
absl-py==2.2.2
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
flatbuffers==25.2.10
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
magic-filter==1.0.12
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.4.3
namex==0.0.9
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==4.25.7
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.3
rich==14.0.0
setuptools==80.7.1
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
